修正 新增權限未帶 course 或 teacher 時, http://localhost:3000/api/v1/Schedules API response沒有資料

+++ DoorWebApp/Controllers/StudentPermissionController.cs
@@ -88,8 +88,8 @@ namespace DoorWebApp.Controllers
                     .Select(sp => new ResStudentPermissionDTO()
                     {
                         Id = sp.Id,
-                        courseId = sp.CourseId,
-                        teacherId = sp.TeacherId,
+                        courseId = (int)sp.CourseId,
+                        teacherId = (int)sp.TeacherId,
                         datefrom = sp.DateFrom,
                         dateto = sp.DateTo,
                         timefrom = sp.TimeFrom,
